+++ test/start-services.ps1
@@ -4,6 +4,7 @@ start "D:\Elasticsearch\elasticsearch-8.14.3\bin\elasticsearch.bat" -workingdire
 start "D:\DynamoDb\Run.bat" -workingdirectory "D:\DynamoDb"
 start powershell {"D:\RavenDB\run.ps1"}
 start "azurite.exe" -workingdirectory "C:\Program Files\Microsoft Visual Studio\2022\Professional\Common7\IDE\Extensions\Microsoft\Azure Storage Emulator" 
+docker run --detach --publish 8081:8081 --publish 1234:1234 --name cosmos-emulator mcr.microsoft.com/cosmosdb/linux/azure-cosmos-emulator:vnext-preview
 
 net start mongodb
 net start mysql80